# FirmaMobileMauiApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FirmaWebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Produkt/{id}": {
      "get": {
        "tags": [
          "Produkt"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Produkt"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Produkt"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Produkt"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Produkt"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Produkt"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/TypProduktu": {
      "get": {
        "tags": [
          "TypProduktu"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypProduktu"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypProduktu"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypProduktu"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TypProduktu"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TypProduktu"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TypProduktu"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TypProduktu"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TypProduktu"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypProduktu"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypProduktu"
                }
              }
            }
          }
        }
      }
    },
    "/api/TypProduktu/{id}": {
      "get": {
        "tags": [
          "TypProduktu"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TypProduktu"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypProduktu"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypProduktu"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TypProduktu"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TypProduktu"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TypProduktu"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TypProduktu"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TypProduktu"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TypTransakcji": {
      "get": {
        "tags": [
          "TypTransakcji"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypTransakcji"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypTransakcji"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypTransakcji"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TypTransakcji"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TypTransakcji"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TypTransakcji"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TypTransakcji"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TypTransakcji"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypTransakcji"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypTransakcji"
                }
              }
            }
          }
        }
      }
    },
    "/api/TypTransakcji/{id}": {
      "get": {
        "tags": [
          "TypTransakcji"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TypTransakcji"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypTransakcji"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypTransakcji"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TypTransakcji"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TypTransakcji"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TypTransakcji"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TypTransakcji"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TypTransakcji"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cena": {
        "required": [
          "waluta",
          "wartosc"
        ],
        "type": "object",
        "properties": {
          "idCeny": {
            "type": "integer",
            "format": "int32"
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "idTypuCeny": {
            "type": "integer",
            "format": "int32"
          },
          "wartosc": {
            "type": "number",
            "format": "double"
          },
          "waluta": {
            "minLength": 1,
            "type": "string"
          },
          "czyAktywna": {
            "type": "boolean"
          },
          "produkt": {
            "$ref": "#/components/schemas/Produkt"
          },
          "typCeny": {
            "$ref": "#/components/schemas/TypCeny"
          }
        },
        "additionalProperties": false
      },
...
      "Kontrahent": {
        "required": [
          "akronim",
          "idKontrahenta",
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "idKontrahenta": {
            "type": "integer",
            "format": "int32"
          },
          "akronim": {
            "minLength": 1,
            "type": "string"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "nip": {
            "type": "string",
            "default": "000-000-00-00",
            "nullable": true
          },
          "idSposobuPlatnosci": {
            "type": "integer",
            "format": "int32"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "sposobPlatnosci": {
            "$ref": "#/components/schemas/SposobPlatnosci"
          },
          "adres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Adres"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PozycjaZZ": {
        "required": [
          "ilosc",
          "rabat"
        ],
        "type": "object",
        "properties": {
          "idPozycjiZleceniaZakupu": {
            "type": "integer",
            "format": "int32"
          },
          "idZleceniaZakupu": {
            "type": "integer",
            "format": "int32"
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "ilosc": {
            "type": "number",
            "format": "double"
          },
          "rabat": {
            "type": "number",
            "format": "double"
          },
          "dataRealizacji": {
            "type": "string",
            "format": "date-time"
          },
          "czyAktywna": {
            "type": "boolean"
          },
          "zlecenieZakupu": {
            "$ref": "#/components/schemas/ZlecenieZakupu"
          },
          "produkt": {
            "$ref": "#/components/schemas/Produkt"
          }
        },
        "additionalProperties": false
      },
...
      "Produkt": {
        "required": [
          "jednostkaMiary",
          "kod",
          "nazwa",
          "producent",
          "stawkaVatSprzedazy",
          "stawkaVatZakupu"
        ],
        "type": "object",
        "properties": {
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "kod": {
            "minLength": 1,
            "type": "string"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "dodatkowaNazwa": {
            "type": "string",
            "nullable": true
          },
          "jednostkaMiary": {
            "minLength": 1,
            "type": "string"
          },
          "symbol": {
            "type": "string",
            "nullable": true
          },
          "swW_PKWiU": {
            "type": "string",
            "nullable": true
          },
          "producent": {
            "minLength": 1,
            "type": "string"
          },
          "idTypu": {
            "type": "integer",
            "format": "int32"
          },
          "fotoURL": {
            "type": "string",
            "nullable": true
          },
          "stawkaVatZakupu": {
            "type": "number",
            "format": "double"
          },
          "stawkaVatSprzedazy": {
            "type": "number",
            "format": "double"
          },
          "odwrotneObciazenie": {
            "type": "boolean"
          },
          "podzielonaPlatnosc": {
            "type": "boolean"
          },
          "stawkaCla": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "stwkaAkcyzy": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "kodPcn": {
            "type": "string",
            "nullable": true
          },
          "kraj": {
            "type": "string",
            "nullable": true
          },
          "czasKompletacji": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "czyAktywna": {
            "type": "boolean"
          },
          "idDostawcy": {
            "type": "integer",
            "format": "int32"
          },
          "typProduktu": {
            "$ref": "#/components/schemas/TypProduktu"
          },
          "dostawca": {
            "$ref": "#/components/schemas/Kontrahent"
          },
          "pozycjaZZ": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PozycjaZZ"
            },
            "nullable": true
          },
          "cena": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cena"
            },
            "nullable": true
          },
          "zlecenieKompletacji": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZlecenieKompletacji"
            },
            "nullable": true
          },
          "produktUbocznyZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduktUbocznyZK"
            },
            "nullable": true
          },
          "skladnikZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkladnikZK"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProduktUbocznyZK": {
        "required": [
          "ilosc"
        ],
        "type": "object",
        "properties": {
          "idPrUbocznego": {
            "type": "integer",
            "format": "int32"
          },
          "idZleceniaKompletacji": {
            "type": "integer",
            "format": "int32"
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ilosc": {
            "type": "number",
            "format": "double"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "zlecenieKompletacji": {
            "$ref": "#/components/schemas/ZlecenieKompletacji"
          },
          "produkt": {
            "$ref": "#/components/schemas/Produkt"
          }
        },
        "additionalProperties": false
      },
      "ProduktyForView": {
        "type": "object",
        "properties": {
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "kod": {
            "type": "string",
            "nullable": true
          },
          "nazwa": {
            "type": "string",
            "nullable": true
          },
          "dodatkowaNazwa": {
            "type": "string",
            "nullable": true
          },
          "jednostkaMiary": {
            "type": "string",
            "nullable": true
          },
          "symbol": {
            "type": "string",
            "nullable": true
          },
          "swW_PKWiU": {
            "type": "string",
            "nullable": true
          },
          "producent": {
            "type": "string",
            "nullable": true
          },
          "idTypu": {
            "type": "integer",
            "format": "int32"
          },
          "typProduktuNazwa": {
            "type": "string",
            "nullable": true
          },
          "fotoURL": {
            "type": "string",
            "nullable": true
          },
          "stawkaVatZakupu": {
            "type": "number",
            "format": "double"
          },
          "stawkaVatSprzedazy": {
            "type": "number",
            "format": "double"
          },
          "odwrotneObciazenie": {
            "type": "boolean"
          },
          "podzielonaPlatnosc": {
            "type": "boolean"
          },
          "stawkaCla": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "stwkaAkcyzy": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "kodPcn": {
            "type": "string",
            "nullable": true
          },
          "kraj": {
            "type": "string",
            "nullable": true
          },
          "czasKompletacji": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "czyAktywna": {
            "type": "boolean"
          },
          "idDostawcy": {
            "type": "integer",
            "format": "int32"
          },
          "dostawcaDane": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SkladnikZK": {
        "required": [
          "ilosc"
        ],
        "type": "object",
        "properties": {
          "idSkladnika": {
            "type": "integer",
            "format": "int32"
          },
          "idZleceniaKompletacji": {
            "type": "integer",
            "format": "int32"
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ilosc": {
            "type": "number",
            "format": "double"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "zlecenieKompletacji": {
            "$ref": "#/components/schemas/ZlecenieKompletacji"
          },
          "produkt": {
            "$ref": "#/components/schemas/Produkt"
          }
        },
        "additionalProperties": false
      },
...
      "SposobPlatnosci": {
        "required": [
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "idSposobuPlatnosci": {
            "type": "integer",
            "format": "int32"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "uwagi": {
            "type": "string",
            "nullable": true
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "kontrahent": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Kontrahent"
            },
            "nullable": true
          },
          "zlecenieZakupu": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZlecenieZakupu"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TypProduktu": {
        "required": [
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "idTypuProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "uwagi": {
            "type": "string",
            "nullable": true
          },
          "fotoURL": {
            "type": "string",
            "nullable": true
          },
          "czySklep": {
            "type": "boolean"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "produkt": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Produkt"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TypTransakcji": {
        "required": [
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "idTypuTransakcji": {
            "type": "integer",
            "format": "int32"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "uwagi": {
            "type": "string",
            "nullable": true
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "zlecenieZakupu": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZlecenieZakupu"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ZlecenieKompletacji": {
        "required": [
          "dataWystawienia",
          "ilosc",
          "oczekiwanaDataRealizacji"
        ],
        "type": "object",
        "properties": {
          "idZleceniaKompletacji": {
            "type": "integer",
            "format": "int32"
          },
          "idKontrahenta": {
            "type": "integer",
            "format": "int32"
          },
          "dataWystawienia": {
            "type": "string",
            "format": "date-time"
          },
          "oczekiwanaDataRealizacji": {
            "type": "string",
            "format": "date-time"
          },
          "potwierdzonaDataRealizacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "ilosc": {
            "type": "number",
            "format": "double"
          },
          "idMontera": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "priorytet": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "czyAktywne": {
            "type": "boolean"
          },
          "czasZlecenia": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kontrahent": {
            "$ref": "#/components/schemas/Kontrahent"
          },
          "monter": {
            "$ref": "#/components/schemas/Monter"
          },
          "produkt": {
            "$ref": "#/components/schemas/Produkt"
          },
          "produktUbocznyZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduktUbocznyZK"
            },
            "nullable": true
          },
          "skladnikZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkladnikZK"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ZlecenieZakupu": {
        "required": [
          "dataWystawienia"
        ],
        "type": "object",
        "properties": {
          "idZleceniaZakupu": {
            "type": "integer",
            "format": "int32"
          },
          "idKontrahenta": {
            "type": "integer",
            "format": "int32"
          },
          "idTransakcji": {
            "type": "integer",
            "format": "int32"
          },
          "idSposobuDostawy": {
            "type": "integer",
            "format": "int32"
          },
          "idRodzajuTransportu": {
            "type": "integer",
            "format": "int32"
          },
          "idSposobuPlatnosci": {
            "type": "integer",
            "format": "int32"
          },
          "nrOferty": {
            "type": "string",
            "nullable": true
          },
          "czyPotwierdzone": {
            "type": "boolean"
          },
          "nrPotwierdzenia": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "totalOrder": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "dataWystawienia": {
            "type": "string",
            "format": "date-time"
          },
          "czyAktywne": {
            "type": "boolean"
          },
          "kontrahent": {
            "$ref": "#/components/schemas/Kontrahent"
          },
          "sposobDostawy": {
            "$ref": "#/components/schemas/SposobDostawy"
          },
          "typTransakcji": {
            "$ref": "#/components/schemas/TypTransakcji"
          },
          "rodzajTransportu": {
            "$ref": "#/components/schemas/RodzajTransportu"
          },
          "sposobPlatnosci": {
            "$ref": "#/components/schemas/SposobPlatnosci"
          },
          "pozycjaZZ": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PozycjaZZ"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}